Remove the notion of remote Cachex instances
