+++ mix.exs
@@ -64,8 +64,7 @@ defmodule Cachex.Mixfile do
       # Production dependencies
       { :eternal, "~> 1.0" },
       # Local dependencies
-      { :benchfella,   "~> 0.3",  optional: true, only: [ :dev, :test ] },
-      { :benchwarmer,  "~> 0.0",  optional: true, only: [ :dev, :test ] },
+      { :benchee,      "~> 0.3",  optional: true, only: [ :dev, :test ] },
       { :credo,        "~> 0.4",  optional: true, only: [ :dev, :test ] },
       { :ex_doc,       "~> 0.13", optional: true, only: [ :dev, :test ] },
       { :excoveralls,  "~> 0.5",  optional: true, only: [ :dev, :test ] },
