+++ test/cachex/take_test.exs
@@ -14,4 +14,33 @@ defmodule Cachex.TakeTest do
     assert(Cachex.take(state_result, "key") == { :missing, nil })
   end
 
+  test "take with existing key", state do
+    set_result = Cachex.set(state.cache, "my_key", "my_value")
+    assert(set_result == { :ok, true })
+
+    get_result = Cachex.get(state.cache, "my_key")
+    assert(get_result == { :ok, "my_value" })
+
+    take_result = Cachex.take(state.cache, "my_key")
+    assert(take_result == { :ok, "my_value" })
+
+    get_result = Cachex.get(state.cache, "my_key")
+    assert(get_result == { :missing, nil })
+  end
+
+  test "take with missing key", state do
+    take_result = Cachex.take(state.cache, "my_key")
+    assert(take_result == { :missing, nil })
+  end
+
+  test "take with expired key", state do
+    set_result = Cachex.set(state.cache, "my_key", "my_value", ttl: 5)
+    assert(set_result == { :ok, true })
+
+    :timer.sleep(10)
+
+    get_result = Cachex.take(state.cache, "my_key")
+    assert(get_result == { :missing, nil })
+  end
+
 end
